Checking in changes in preparation for group behaviors - pmpl merge. This check in is specifically aimed at the redesign of the Cfg_2D class to be true 2DOF. Cfg_2D_withRot has rotation but remains in the 2D workspace. Also changes to the environment are to allow for actual 2D workspace definitions.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@2847 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/MPProblem/Boundary.cpp
@@ -98,13 +98,13 @@ operator==(const BoundingBox& bb) const
          (dofs == bb.dofs);
 }
 
-unsigned int 
+int 
 BoundingBox::
 GetDOFs() const {
   return dofs;//bounding_box.size();
 }
 
-unsigned int 
+int 
 BoundingBox::
 GetPosDOFs() const {
   return pos_dofs;//bounding_box.size();
@@ -172,7 +172,7 @@ BoundingBox
 BoundingBox::
 GetCombination(BoundingBox &o_bounding_box) {
   BoundingBox combination(*this);
-  for (size_t par=0; par < o_bounding_box.GetDOFs() && par < combination.GetDOFs(); par++) {
+  for (int par=0; par < o_bounding_box.GetDOFs() && par < combination.GetDOFs(); par++) {
     combination.SetParameter(par,
 			     min(combination.bounding_box[par].first,
 				 o_bounding_box.bounding_box[par].first),
